Revert "refactor: restructure utils and events as PR#43"
This reverts commit b53ce2873c617ccbe9f1e815e37ddbe7b8896f3b.
