--- pos_next/utils/__init__.py
@@ -1,1 +1,0 @@
-from .version import get_build_version, get_app_version
